+++ src/emma_policy/inference/model_wrapper/simbot_nlu_output_processor.py
@@ -1,4 +1,6 @@
-from typing import Any, Optional
+from typing import Optional
+
+from emma_common.datamodels import EmmaExtractedFeatures
 
 from emma_policy.datamodules.simbot_nlu_dataset import SimBotNLUIntents
 
@@ -10,13 +12,13 @@ class SimBotNLUPredictionProcessor:
         self.valid_action_types = valid_action_types
         self._default_prediction = default_prediction
 
-    def __call__(self, prediction: str, frame_features: list[dict[str, Any]]) -> str:
+    def __call__(self, prediction: str, frame_features: list[EmmaExtractedFeatures]) -> str:
         """Process the prediction."""
-        new_prediction = self._overwrite_the_nlu_prediction(prediction)
+        object_name = self._get_target_object(prediction)
+        new_prediction = self._overwrite_the_nlu_prediction(prediction, object_name)
         if new_prediction != prediction:
             return new_prediction
 
-        object_name = self._get_target_object(prediction)
         if object_name is None:
             return prediction
         if prediction.startswith(SimBotNLUIntents.act_no_match.value):
@@ -37,14 +39,14 @@ class SimBotNLUPredictionProcessor:
         prediction_type = prediction.split(" ")[0]
         return prediction_type in self.valid_action_types
 
-    def _overwrite_the_nlu_prediction(self, prediction: str) -> str:
+    def _overwrite_the_nlu_prediction(self, prediction: str, object_name: Optional[str]) -> str:
         """Check if the predicted NLU output needs to be overwritten."""
         # If the predicted prediction is not valid return the default prediction
         if not self._prediction_type_is_valid(prediction):
             return self._default_prediction
-        # For search intents only return <search>
+        # For search intents only return <search> object_name
         if prediction.startswith(SimBotNLUIntents.search.value):
-            return SimBotNLUIntents.search.value
+            return f"{SimBotNLUIntents.search.value} {self._get_target_object(prediction)}"
         # For act one_match intents only return <act><one_match>
         if prediction.startswith(SimBotNLUIntents.act_one_match.value):
             return SimBotNLUIntents.act_one_match.value
@@ -55,15 +57,17 @@ class SimBotNLUPredictionProcessor:
         split_parts = prediction.split(" ")
         return " ".join(split_parts[1:]) if len(split_parts) > 1 else None
 
-    def _get_detected_objects(self, frame_features: list[dict[str, Any]]) -> Optional[list[str]]:
+    def _get_detected_objects(
+        self, frame_features: list[EmmaExtractedFeatures]
+    ) -> Optional[list[str]]:
         """Get a list of class labels fro the detected objects."""
-        class_labels = frame_features[0].get("class_labels", None)
+        class_labels = frame_features[0].class_labels
         if class_labels is not None:
             class_labels = [label.lower() for label in class_labels]
         return class_labels
 
     def _rule_based_ambiguity_check(
-        self, prediction: str, frame_features: list[dict[str, Any]], object_name: str
+        self, prediction: str, frame_features: list[EmmaExtractedFeatures], object_name: str
     ) -> str:
         """Change too_many_matches prediction if there is one detected object."""
         # For now, overwrite the NLU only if there are no multiples in front of you
@@ -79,7 +83,7 @@ class SimBotNLUPredictionProcessor:
         return prediction
 
     def _special_robotic_arm_button_case(
-        self, prediction: str, frame_features: list[dict[str, Any]]
+        self, prediction: str, frame_features: list[EmmaExtractedFeatures]
     ) -> str:
         class_labels = self._get_detected_objects(frame_features)
         if class_labels is None: